Disabled Android Studio's "instant run" and Retrofit now works.Changes some models to match the response structure and alters the Rest client for use with RxJava.

+++ app/src/main/java/com/cedarbarkgrooming/model/weather/Coord.java
@@ -5,7 +5,7 @@ package com.cedarbarkgrooming.model.weather;
  */
 public class Coord {
 
-    public int lon;
-    public int lat;
+    public double lon;
+    public double lat;
 
 }
